fix: dtype=float is not recommended, use torch.float
Surprising behaviour: float is actually 64 bits, not 32, because Python's float
is double, and has been for a long time, in the same high precision philosophy
that git Python int to be arbitrary size, not a machine word limited type like
int32 or int64.

For a torch.tensor constructor, you're supposed to use torch types.
dtype=torch.float is the expected 32 bit type, but it's even better to be
explicit with dtype=torch.float32 so the code can be self-documenting.

Also adjust the test cases for 32 bit precision stats calculation instead of 64.

+++ everyvoice/preprocessor/preprocessor.py
@@ -447,13 +447,15 @@ class Preprocessor:
                 print("Gathering character length values")
                 for line in filelist:
                     char_length_data = torch.tensor(
-                        [len(line["characters"])], dtype=float
+                        [len(line["characters"])], dtype=torch.float32
                     )
                     char_length_scaler.append(char_length_data)
             if phone_length:
                 phone_length_scaler = Scaler()
                 for line in tqdm(filelist, desc="Gathering phone length values"):
-                    phone_length_data = torch.tensor([len(line["phones"])], dtype=float)
+                    phone_length_data = torch.tensor(
+                        [len(line["phones"])], dtype=torch.float32
+                    )
                     phone_length_scaler.append(phone_length_data)
 
         return energy_scaler, pitch_scaler, char_length_scaler, phone_length_scaler
